+++ lib/pleroma/web/activity_pub/utils.ex
@@ -822,18 +822,31 @@ defmodule Pleroma.Web.ActivityPub.Utils do
     reports = get_reports_by_status_id(activity["id"])
     max_date = Enum.max_by(reports, &NaiveDateTime.from_iso8601!(&1.data["published"]))
     actors = Enum.map(reports, & &1.user_actor)
+    {deleted, status} = get_status_data(activity)
 
     %{
       date: max_date.data["published"],
       account: activity["actor"],
-      status: %{
-        id: activity["id"],
-        content: activity["content"],
-        published: activity["published"]
-      },
+      status: status,
+      status_deleted: deleted,
       actors: Enum.uniq(actors),
       reports: reports
     }
+  end
+
+  defp get_status_data(activity) do
+    case Activity.get_by_ap_id(activity["id"]) do
+      %Activity{} = act ->
+        {false, act}
+
+      _ ->
+        {true,
+         %{
+           id: activity["id"],
+           content: activity["content"],
+           published: activity["published"]
+         }}
+    end
   end
 
   def get_reports_by_status_id(ap_id) do